+++ Project/GameObject/Player/PlayerConfig.h
@@ -41,7 +41,7 @@ const Vector3 kMaxRotSpeed = { 1.0f, 1.0f, 1.0f };
 // 自機の最大回避速度
 const Vector2 kMaxEvasionMoveSpeed = { 4.0f, 4.0f };
 // 回避時の回転数[2π * n回]
-const Vector3 kMaxEvasionRotNum = { 0.0f, 0.0f ,(2.0f * std::numbers::pi) * 2.0f };
+const Vector3 kMaxEvasionRotNum = { 0.0f, 0.0f ,(2.0f * (float)std::numbers::pi) * 2.0f };
 
 // 移動限界座標
 const Vector3 kMoveLimit = { 35.0f, 17.0f, 5.0f };
